requirements.txt
torch>=2.5.0
numpy==1.26.4
pyyaml==6.0.1
pandas==2.2.0
matplotlib==3.8.3
tensorboard>=2.17.0
python-dotenv==1.0.0
Flask>=3.1.0
requests>=2.32.4
beautifulsoup4==4.12.3
lxml==4.9.4
PyQt5==5.15.10
PyQtWebEngine==5.15.7
transformers>=4.47.0
tokenizers>=0.21.0
faiss-cpu==1.8.0.post1
onnx==1.17.0
protobuf>=5.26.0
